ocr_service/requirements.txt
numpy==1.26.4
opencv-python==4.8.1.78
torch
torchvision
easyocr
pillow
groq
python-dotenv
